pyproject.toml
[tool.poetry]
name = "user-center-backend"
version = "0.1.0"
description = ""
authors = ["Johan Guse <johanguse@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "3.11.*"
charset-normalizer = "^3.1.0"  # Needed cause of a bug in poetry
urllib3 = "^2.0"  # Needed cause of a bug in poetry
fastapi = {extras = ["standard"], version = "^0.115.7"}
fastapi-users = {extras = ["oauth", "sqlalchemy"], version = "^14.0.1"}
fastapi-pagination = {extras = ["sqlalchemy"], version = "^0.12.34"}
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.9"
pydantic = {extras = ["email"], version = "^2.8.2"}
alembic = "^1.14.1"
uvicorn = "^0.34.0"
pyjwt = {extras = ["crypto"], version = "^2.9.0"}
passlib = {extras = ["argon2"], version = "^1.7.4"}
python-dotenv = "^1.0.1"
pydantic-settings = "^2.4.0"
python-multipart = "^0.0.20"
pandas = "^2.2.2"
chardet = "^5.2.0"
celery = "^5.4.0"
resend = "^2.4.0"
asyncpg = "^0.30.0"
boto3 = "^1.36.4"
httpx-oauth = "^0.16.1"
sqlalchemy-utils = "^0.41.2"
loguru = "^0.7.3"
python-json-logger = "^3.2.1"
ua-parser = "^1.0.0"
user-agents = "^2.2.0"
python-terraform = "^0.10.1"
cachetools = "^5.5.1"
stripe = "^11.4.1"
cryptography = "^44.0.0"
python-magic = "^0.4.27"
langchain = "^0.3.15"
llama-index = "^0.12.14"
openai = "^1.60.1"
llama-index-core = "^0.12.14"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"
ruff = "^0.5.7"
httpx = "^0.27.0"
testcontainers = {version = "^4.8.0", extras = ["postgresql"]}
factory-boy = "^3.3.0"
pytest-asyncio = "^0.23.5"
aiosqlite = "^0.19.0"
asyncpg = "^0.30.0"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']
ignore = ["E203", "E501", "PLC0415"]

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.ruff.lint.per-file-ignores]
"src/activity_log/service.py" = ["PLC0415"]

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check src && ruff check src --diff'
format = 'ruff check src --fix && ruff format src'
run = 'uvicorn src.main:app --reload --port 8001'
dev = 'uvicorn src.main:app --reload --port 8001'
pre_test = 'task lint'
test = 'pytest -s -x --cov=src -vv'
post_test = 'coverage run -m pytest --cov=src'
test-auth = 'pytest tests/test_auth.py -v'
test-watch = 'pytest -xvs --watch'
migrations = 'alembic revision --autogenerate -m'
migrate = 'alembic upgrade head'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
